ARM: dts: samsung: exynos4210: set sane thermal trip points
The previous first CPU thermal trip point at 85 degrees celcius was
insane. At such a high temperature, with i9100, the phone gets
disturbingly hot to the touch.

So, set much lower thermal limits, in line with the limits used in the
downstream kernel for i9100 [1].

While at it, remove the override for cpu_alert2 from exynos4210.dtsi,
since the temperature set is the exact same as the one set within
exynos4-cpu-thermal.dtsi.

TODO - Should this be set exclusively for i9100, or really for the
Exynos 4210 entirely? Unsure. This is easier for personal use in any
case.

The front camera even fails to initialize, seemingly due to high
temperatures. When the phone is cooled down, the camera initializes
properly. But the failures start occurring at a point lower than even
the new limits, anyway, so that's odd. Issue with the front camera
firmware being used perhaps?

[1] https://github.com/LineageOS/android_kernel_samsung_smdk4412/blob/b7ffe7f2aea2391737cdeac2a33217ee0ea4f2ba/arch/arm/mach-exynos/mach-u1.c#L7503

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>