usb: gadget: u_ether: Revert "Replace netif_stop_queue with netif_device_detach"
This reverts commit f49449fbc21e7e9550a5203902d69c8ae7dfd918.

It seems to make USB gadget ethernet interfaces unusable; A gadget
device, presenting an NCM device class, has its USB ethernet interface
claimed on it as being disabled, seemingly after being enabled and then
disabled again once, and attempts to enable it thereafter fail, while
another device that the gadget device is connected to has an USB
ethernet interface presented to it but absolutely all communication
through it fails.

Hopefully, those issues since that commit will be ironed out later, or
it reverted if it happens that it's not even all that valid in the first
place, which I couldn't really tell myself. But for now, this revert
will do.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>